pyproject.toml
[tool.poetry]
name = "keycloak-docker"
version = "0.1.0"
description = ""
authors = ["Max Pfeiffer <max@maxpfeiffer.ch>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "3.11.*"
click = "8.1.7"
python-on-whales = "0.71.0"

[tool.poetry.group.dev.dependencies]
docker-image-py = "0.1.12"
furl = "2.1.3"
httpx = "0.27.0"
pre-commit = "3.7.0"
pytest = "8.1.1"
pytest-cov = "5.0.0"
testcontainers = "4.4.0"
semver = "3.0.2"
trustme = "1.1.0"
bcrypt = "4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [".venv"]

[tool.ruff.lint]
select = [
    "F", # Pyflakes
    "E", # pycodestyle
    "W", # pycodestyle
    "I", # isort
    "D",   # pydocstyle
    "UP",  # pyupgrade
    "ASYNC", # flake8-async
    "RUF",  # Ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "pep257"

# https://docs.pytest.org/en/latest/reference/customize.html
[tool.pytest.ini_options]
testpaths = [
    "tests",
]
